Feature/notificationfix (#22)
* Fix race condition in notifications

While the cluster is being deleted, it still exists for a short while.
If the notifications job gets run during this time, the history would
have been removed so duplicate alerts would get sent out.

We therefore keep the cluster history and rely on the notify job to do
the histiry cleanup as it currently does anyway,

* remove unused notification history import and methods

---------

Co-authored-by: Mark Dastmalchi-Round <gitlab@markround.com>